Move kotlin code to its own package
